+++ example/lib/main.dart
@@ -467,7 +467,6 @@ class _MyHomePageState extends State<MyHomePage> {
 
     LineChart lineChart = new LineChart(
       painter: new LineChartPainter(lineChartContainer: lineChartContainer),
-      container: lineChartContainer,
     );
 
     VerticalBarChartContainer verticalBarChartContainer =
@@ -480,7 +479,6 @@ class _MyHomePageState extends State<MyHomePage> {
     VerticalBarChart verticalBarChart = new VerticalBarChart(
       painter: new VerticalBarChartPainter(
           verticalBarChartContainer: verticalBarChartContainer),
-      container: verticalBarChartContainer,
     );
 
     // [MyHomePage] extends [StatefulWidget].